--- Nations/src/main/java/net/pixlies/nations/entities/chunk/NationChunkType.java
@@ -1,4 +1,0 @@
-package net.pixlies.nations.entities.chunk;
-
-public enum NationChunkType {
-}
