+++ Nations/src/main/java/net/pixlies/nations/nations/chunk/NationChunkType.java
@@ -1,0 +1,4 @@
+package net.pixlies.nations.nations.chunk;
+
+public enum NationChunkType {
+}
